fix：懒加载tooltip并修改e2e
